fixed bug with duplicated calls
