+++ client/src/ResourcesPortalContext.js
@@ -8,6 +8,7 @@ export const ResourcesPortalContextProvider = ({ children }) => {
   const [user, setUser] = useLocalStorage('user', undefined)
   const [token, setToken] = useLocalStorage('token', '')
   const [alertsQueues, setAlertsQueues] = React.useState({})
+  const [notifications, setNotifications] = React.useState()
 
   return (
     <ResourcesPortalContext.Provider
@@ -19,7 +20,9 @@ export const ResourcesPortalContextProvider = ({ children }) => {
         token,
         setToken,
         alertsQueues,
-        setAlertsQueues
+        setAlertsQueues,
+        notifications,
+        setNotifications
       }}
     >
       {children}
